requirements.txt
amqp==1.4.9
anyjson==0.3.3
backports.shutil-get-terminal-size==1.0.0
billiard==3.3.0.23
celery==3.1.0
colorama==0.3.7
decorator==4.0.10
Django==1.9
django-celery==3.1.17
django-pipeline==1.6.8
django-pipeline-browserify==0.4.1
django-webpack-loader==0.3.3
djangorestframework==3.4.6
gunicorn==19.6.0
ipython==5.1.0
ipython-genutils==0.1.0
kombu==3.0.37
pathlib2==2.1.0
pickleshare==0.7.4
prompt-toolkit==1.0.7
Pygments==2.1.3
pytz==2016.10
redis==2.10.5
simplegeneric==0.8.1
simplejson==3.8.2
six==1.10.0
traitlets==4.2.2
vine==1.1.3
wcwidth==0.1.7
whitenoise==3.2.3
win-unicode-console==0.5
yahoo-finance==1.2.1
